% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_french}
\alias{download_french}
\title{Download file from French's website}
\usage{
download_french(file, exdir = ".")
}
\arguments{
\item{file}{File name on http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html.}

\item{path}{Path where extracted file is saved.}

\item{overwrite}{If \code{TRUE}, overwrite existing files.}
}
\description{
Download file from French's website
}
